+++ frontend/src/QueryKeys/QueryKeys.js
@@ -1,0 +1,5 @@
+const QueryKeys = {
+  DashboardData: 'dashboardData',
+};
+
+export default QueryKeys;
